Modifying what gets stored in the session cookie.
Instead of storing the user account, we'll now *only* store the user's account
HREF. This is a tiny bit more secure, but more importantly, less resource
intensive. Since the cookie is now only a string, it's much smaller and requires
much less bandwidth to transmit back to the server on subsequent requests.