Fixing Stormpath middleware requirement handling.
Previously, the way all our middlwares worked was as follows:

- A user would be authenticated from either a cookie, or an API request (HTTP
  Authorization header).
- A developer would then include a Stormpath middleware like
  `stormpath.loginRequired`, which tells the developer that a specific route
  should only let authenticated web users in.
- But then, an API user could *also* get through that barrier.

The way this works now is a bit different, we're actually enforcing a specific
type of auth for the required middlewares.

So `stormpath.apiAuthenticationRequired`, for instance, will *only* let through
users who have successfully authenticated via the HTTP Authorization header.

This will be updated in the future to use a cleaner approach.