Fixes to the readme
The way it was written, we were telling people that single-page applications
were meant to be used with the `api` option.  This is not correct, SPAs
need the `website` option.  The `api` option enables or OAuth workflows,
nothing else.

I still need to file a Discuss issue so that we can discuss how we
feel about these option profiles.  But in the meantime I want to 
make these updates to the README so that we reduce confusion