Update to latest config, remove website and api options
This commit brings the config.yaml up-to-date with the latest copy from the framework spec.

As such, several of the form parsing files needed modifications in order to find the field and filed order definitions in their new locations, under the `form` key.

With this commit I am also removing the `website` and `api` options, as intended for the 3.0 release.  The features that were enabled by these options are now enabled by default, so these options are no longer necessary.