Updates to the Setup and Configuration documentation
Previously, the steps required to get started with this module were split between the “Setup” page and the “Configuration” page.

This PR fixes that strangeness.  The “Setup” page is now quickstart-like, and the “Configuration” page has more in-depth information about how to configure the library, and what the defaults are.